requirements.txt
crewai
crewai-tools
uv
ruff
pytest
python-dotenv
langchain-openai
